requirements.txt
gradio==4.44.1
langchain==0.3.24
langchain-community==0.3.22
langchain-core==0.3.55
langchain-openai==0.3.14
langchain-chroma==0.2.3
openai==1.75.0
python-dotenv==1.0.0
pandas==2.2.3
numpy==1.26.4
